Update Source/System/GraphicalPrimitive.cpp
implicit cast

+++ Source/System/GraphicalPrimitive.cpp
@@ -509,7 +509,7 @@ void BitmapPrimitive::Draw(BITMAP* drawScreen, const Vector& targetPos) {
 	if (m_Scale == 1.0f) {
 		draw_sprite(bitmapToDraw, m_Bitmap, 0, 0);
 	} else {
-		stretch_sprite(bitmapToDraw, m_Bitmap, 0, 0, (int)(m_Bitmap->w * m_Scale), (int)(m_Bitmap->h * m_Scale));
+		stretch_sprite(bitmapToDraw, m_Bitmap, 0, 0, m_Bitmap->w * m_Scale, m_Bitmap->h * m_Scale);
 	}
 
 	if (m_HFlipped || m_VFlipped) {
